refactor(controlplane): relocate infra-IP helpers to infra_lookup.go (KD-3b prep)
PR-8 prep, commit 1 of the decomposition in
.claude/decisions/0001-pr8-providerid-elimination.md.

Move three unstructured-client helpers out of the 2400-line
kairoscontrolplane_controller.go and into a new file:

  getNodeIP                  CAPD/CAPV/CAPK node-IP resolver
  extractIPFromUnstructured  status.addresses / status.network / vmIp probe
  getKubevirtVMIIP           VMI status.interfaces[].ipAddress probe

These helpers cannot be deleted alongside ensureProviderIDOnNodes in the
later commits of this PR — they have surviving callers in updateClusterStatus
and the LB-endpoint reconciler. Extracting them now keeps those callers
working while making the upcoming delete commits a clean cut against the
original controller file.

Pure cut/paste — no logic change. Drop the now-unused schema import from
kairoscontrolplane_controller.go (its only references were inside the
moved functions). Build + unit tests + envtest pass.
